requirements.txt
langchain==1.0.0
langchain-groq==1.0.0
fastapi==0.111.0
uvicorn==0.30.0
prometheus-client==0.20.0
python-dotenv==1.0.1
qdrant-client<2.0.0
pgvector==0.3.3
psycopg2-binary==2.9.9
langchain-community==0.4
sentence-transformers==5.1.2
langchain-core==1.0.1
langchain-huggingface==1.0.0
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
langchain-classic==1.0.0
boto3